backend/requirements.txt
Django==6.0.2
django-ninja==1.5.3
django-cors-headers==4.9.0